+++ packages/eslint-plugin/tests/rules/no-unsafe-return.test.ts
@@ -108,6 +108,17 @@ function foo(): Set<number> {
         return new Map();
       }
     `,
+    // https://github.com/typescript-eslint/typescript-eslint/issues/3549
+    `
+      function foo(): any {
+        return [] as any[];
+      }
+    `,
+    `
+      function foo(): unknown {
+        return [] as any[];
+      }
+    `,
   ],
   invalid: [
     {